Remove StatusListTokenIntegrityValidator class
