+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/StatusListToken.kt
@@ -1,10 +1,9 @@
 package at.asitplus.wallet.lib.data
 
 import at.asitplus.KmmResult
+import at.asitplus.catching
 import at.asitplus.signum.indispensable.cosef.CoseSigned
 import at.asitplus.signum.indispensable.josef.JwsSigned
-import at.asitplus.wallet.lib.agent.StatusListCwtIntegrityValidator
-import at.asitplus.wallet.lib.agent.StatusListJwtIntegrityValidator
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignature
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureFun
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListInfo
@@ -45,17 +44,52 @@ sealed interface StatusListToken {
         override val payload: StatusListTokenPayload
             get() = value.payload
 
+        /**
+         * Validate the Status List Token:
+         *
+         * Validate the Status List Token by following the rules defined in section 7.2 of [RFC7519]
+         * for JWTs and section 7.2 of [RFC8392] for CWTs
+         *
+         * Check for the existence of the required claims as defined in Section 5.1 and Section 5.2
+         * depending on token type.
+         */
         suspend fun validate(
             verifyJwsObject: VerifyJwsObjectFun = VerifyJwsObject(),
             statusListInfo: StatusListInfo,
             isInstantInThePast: (Instant) -> Boolean,
-        ): KmmResult<StatusListTokenPayload> = StatusListTokenValidator.validateStatusListToken(
-            statusListToken = this,
-            statusListTokenResolvedAt = resolvedAt,
-            integrityValidator = StatusListJwtIntegrityValidator(verifyJwsObject = verifyJwsObject),
-            statusListInfo = statusListInfo,
-            isInstantInThePast = isInstantInThePast,
-        )
+        ): KmmResult<StatusListTokenPayload> =
+            validateIntegrity(verifyJwsObject, this).transform { payload ->
+                StatusListTokenValidator.validateStatusListTokenPayloadClaims(
+                    statusListTokenPayload = payload,
+                    statusListTokenResolvedAt = resolvedAt,
+                    statusListInfo = statusListInfo,
+                    isInstantInThePast = isInstantInThePast,
+                )
+            }
+
+        /**
+         * Validate the integrity of a status list jwt
+         */
+        private suspend fun validateIntegrity(
+            verifyJwsObject: VerifyJwsObjectFun,
+            statusListToken: StatusListJwt
+        ): KmmResult<StatusListTokenPayload> =
+            catching {
+                val jwsSigned = statusListToken.value
+                verifyJwsObject(jwsSigned).getOrElse {
+                    throw IllegalStateException(it)
+                }
+                val type = jwsSigned.header.type?.lowercase()
+                    ?: throw IllegalArgumentException("Invalid type header")
+                val validTypes = listOf(
+                    MediaTypes.STATUSLIST_JWT.lowercase(),
+                    MediaTypes.Application.STATUSLIST_JWT.lowercase()
+                )
+                if (type !in validTypes) {
+                    throw IllegalArgumentException("Invalid type header: $type")
+                }
+                jwsSigned.payload
+            }
     }
 
     data class StatusListCwt(
@@ -65,16 +99,48 @@ sealed interface StatusListToken {
         override val payload: StatusListTokenPayload
             get() = value.payload ?: throw IllegalStateException("Payload not found.")
 
+        /**
+         * Validate the Status List Token:
+         *
+         * Validate the Status List Token by following the rules defined in section 7.2 of [RFC7519]
+         * for JWTs and section 7.2 of [RFC8392] for CWTs
+         *
+         * Check for the existence of the required claims as defined in Section 5.1 and Section 5.2
+         * depending on token type.
+         */
         suspend fun validate(
             verifyCoseSignature: VerifyCoseSignatureFun<StatusListTokenPayload> = VerifyCoseSignature(),
             statusListInfo: StatusListInfo,
             isInstantInThePast: (Instant) -> Boolean,
-        ): KmmResult<StatusListTokenPayload> = StatusListTokenValidator.validateStatusListToken(
-            statusListToken = this,
-            statusListTokenResolvedAt = resolvedAt,
-            integrityValidator = StatusListCwtIntegrityValidator(verifyCoseSignature = verifyCoseSignature),
-            statusListInfo = statusListInfo,
-            isInstantInThePast = isInstantInThePast,
-        )
+        ): KmmResult<StatusListTokenPayload> =
+            validateIntegrity(verifyCoseSignature, this).transform { payload ->
+                StatusListTokenValidator.validateStatusListTokenPayloadClaims(
+                    statusListTokenPayload = payload,
+                    statusListTokenResolvedAt = resolvedAt,
+                    statusListInfo = statusListInfo,
+                    isInstantInThePast = isInstantInThePast,
+                )
+            }
+
+        /**
+         * Validate the integrity of a status list cwt
+         */
+        private suspend fun validateIntegrity(
+            verifyCoseSignature: VerifyCoseSignatureFun<StatusListTokenPayload>,
+            statusListToken: StatusListCwt
+        ): KmmResult<StatusListTokenPayload> =
+            catching {
+                val coseStatus = statusListToken.value
+                verifyCoseSignature(coseStatus, byteArrayOf(), null).getOrElse {
+                    throw IllegalStateException("Invalid Signature.", it)
+                }
+                val type = coseStatus.protectedHeader.type?.lowercase()
+                    ?: throw IllegalArgumentException("Invalid type header")
+                if (type != MediaTypes.Application.STATUSLIST_CWT.lowercase()) {
+                    throw IllegalArgumentException("Invalid type header: $type")
+                }
+                coseStatus.payload
+                    ?: throw IllegalStateException("Status list token payload not found.")
+            }
     }
 }
